pyproject.toml
[project]
name = "dataanalysis"
version = "0.1.0"
description = "AI agent crew for analyzing unknown datasets"
readme = "README.md"
requires-python = "==3.12"
dependencies = [
    "crewai[tools]>=0.28.0",
    "e2b-code-interpreter>=0.14.1",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "ruff>=0.1.5",
    "mypy>=1.5.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
lint.select = ["E", "F", "I", "B"]
lint.ignore = ["E501"]  # Ignore line too long errors

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true